pyproject.toml
[project]
name = "goit_pythonweb_hw_08"
version = "0.1.0"
description = ""
authors = [
    {name = "Andrii.Rohovenko",email = "andrii.rohovenko@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "sqlalchemy (>=2.0.43,<3.0.0)",
    "alembic (>=1.16.5,<2.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "uvicorn[standard] (>=0.35.0,<0.36.0)",
    "fastapi[standard] (>=0.116.1,<0.117.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "asyncio (>=4.0.0,<5.0.0)",
    "pydantic-settings (>=2.10.1,<3.0.0)",   
]

[tool.poetry]
package-mode = true

packages = [
    { include = "src" }
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
dev = "src.run:dev"
start-prod = "src.run:prod"
